Move encryptor to shared package
