+++ .changeset/chatty-comics-behave.md
@@ -1,0 +1,5 @@
+---
+"saleor-app-payment-stripe": patch
+---
+
+Move Encryptor to @saleor/apps-shared
