+++ .changeset/long-gifts-begin.md
@@ -1,0 +1,5 @@
+---
+"@saleor/apps-shared": minor
+---
+
+Add Encryptor from Stripe
