+++ apps/stripe/src/__tests__/mocks/mock-encryptor.ts
@@ -1,4 +1,5 @@
+import { Encryptor } from "@saleor/apps-shared/encryptor";
+
 import { mockedEncryptionKey } from "@/__tests__/mocks/mocked-encryption-key";
-import { Encryptor } from "@/lib/encryptor";
 
 export const mockEncryptor = new Encryptor(mockedEncryptionKey);
