+++ packages/shared/package.json
@@ -15,7 +15,8 @@
     "./race": "./src/race.ts",
     "./saleor-version-compatibility-validator": "./src/saleor-version-compatibility-validator.ts",
     "./theme-synchronizer": "./src/theme-synchronizer.tsx",
-    "./use-dashboard-notification": "./src/use-dashboard-notification.ts"
+    "./use-dashboard-notification": "./src/use-dashboard-notification.ts",
+    "./encryptor": "./src/encryptor.ts"
   },
   "scripts": {
     "check-types": "tsc",
